# MM-60706: Implement a free-list InvalidateProfilesInChannelCacheByUser