--- a/server/channels/store/localcachelayer/layer.go
+++ b/server/channels/store/localcachelayer/layer.go
@@ -12,6 +12,7 @@ import (
 	"github.com/mattermost/mattermost/server/v8/channels/store"
 	"github.com/mattermost/mattermost/server/v8/einterfaces"
 	"github.com/mattermost/mattermost/server/v8/platform/services/cache"
+	"github.com/mattermost/mattermost/server/v8/platform/shared/dequeue"
 )
 
 const (
@@ -353,6 +354,7 @@ func NewLocalCacheLayer(baseStore store.Store, metrics einterfaces.MetricsInterf
 		UserStore:                     baseStore.User(),
 		rootStore:                     &localCacheStore,
 		userProfileByIdsInvalidations: make(map[string]bool),
+		userMapFreeList:               dequeue.New[model.UserMap](),
 	}
 
 	// Teams