We implement a free-list of model.UserMap objects just for
InvalidateProfilesInChannelCacheByUser. From looking at heap profiles
at very high scales, it is clear that this method is called very often
and generate a lot of garbage primarily from generating a lot of
model.UserMap objects. Especially, the 3 map fields Props, NotifyProps
and Timezone.

This is due to the reason we have to scan the entire keyspace
of the cache to find out which users are part of the channels
and then delete them. But the allocated userMap objects
go out of scope as soon as the method is finished. This creates
extremely short-lived objects at a very high frequency.

We do not choose a sync.Pool because it is not suitable
for this purpose. Objects in sync.Pool are liable to be GC'd any time
and it allows for dynamically growing and shrinking of the pool.

For our purposes, we simply copy over the stdlib's container/list
implementation and generi-fy it and use a simple dequeue to store
the objects.

The current benchmark was artificially made to have just 3 channels,
but in production there would be hundreds and thousands of channels
which would lead to better results.

```
goos: linux
goarch: amd64
pkg: github.com/mattermost/mattermost/server/v8/channels/store/localcachelayer
cpu: Intel(R) Core(TM) i5-8265U CPU @ 1.60GHz
                                         │ master.txt  │              new.txt               │
                                         │   sec/op    │   sec/op     vs base               │
InvalidateProfilesInChannelCacheByUser-8   5.378µ ± 1%   5.694µ ± 1%  +5.88% (p=0.000 n=10)

                                         │  master.txt  │               new.txt                │
                                         │     B/op     │     B/op      vs base                │
InvalidateProfilesInChannelCacheByUser-8   5.984Ki ± 0%   5.328Ki ± 0%  -10.97% (p=0.000 n=10)

                                         │ master.txt │              new.txt              │
                                         │ allocs/op  │ allocs/op   vs base               │
InvalidateProfilesInChannelCacheByUser-8   68.00 ± 0%   65.00 ± 0%  -4.41% (p=0.000 n=10)
```

Another with 8 channels

```
goos: linux
goarch: amd64
pkg: github.com/mattermost/mattermost/server/v8/channels/store/localcachelayer
cpu: Intel(R) Core(TM) i5-8265U CPU @ 1.60GHz
                                         │ master_8.txt │              pr_8.txt              │
                                         │    sec/op    │   sec/op     vs base               │
InvalidateProfilesInChannelCacheByUser-8    12.28µ ± 4%   12.77µ ± 1%  +4.01% (p=0.011 n=10)

                                         │ master_8.txt │               pr_8.txt               │
                                         │     B/op     │     B/op      vs base                │
InvalidateProfilesInChannelCacheByUser-8   13.87Ki ± 0%   12.34Ki ± 0%  -11.04% (p=0.000 n=10)

                                         │ master_8.txt │             pr_8.txt              │
                                         │  allocs/op   │ allocs/op   vs base               │
InvalidateProfilesInChannelCacheByUser-8     152.0 ± 0%   145.0 ± 0%  -4.61% (p=0.000 n=10)
```

https://mattermost.atlassian.net/browse/MM-60706

```release-note
NONE
```